Converted some comments to doc strings. Removed deprecated apply() calls. Changed quicktimer to use datetime class.
Added a static method to Win32TimeZone called GetTimeZones which returns the list of the available time zone names (like you see when configuring a time zone).

inherited the time period calculation functions from the deprecated log module.

created a TimestampFileHandler which is a Python 2.3 logging handler for logging to files with an optional timestamp.  This functionality was also taken from the log module.

Finally, the LogFileWrapper was also stolen from the log module.  Seems like the Python writers didn't consider everything when they wrote the generic logging module.